Refactor Bookmark code Fixes #195
