+++ src/main/java/de/embl/cba/mobie/bookmark/write/NameAndFileLocation.java
@@ -1,0 +1,9 @@
+package de.embl.cba.mobie.bookmark.write;
+
+import de.embl.cba.tables.FileUtils;
+
+public class NameAndFileLocation
+{
+	public String name;
+	public FileUtils.FileLocation location;
+}
